requirements.txt
pyaudio
numpy
openai
openai[realtime]
sounddevice
pydub
pvporcupine
pveagle
pvrecorder
requests
bs4
nextion
deezer-downloader
pyjwt
cryptography
dotenv